html/semantics/embedded-content-0/media-elements/src_reflects_no_value.html

<!doctype html>
<html>
 <head>
  <title>{audio,video}.src - empty if no source</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-media-src">spec reference</a></p>
  <audio id="audio">
  </audio>
  <video id="video">
  </video>
  <div id="log"></div>
  <script>
test(function() {
 assert_equals(
  document.getElementById("audio").src,
  "",
  "audioElement.src should be empty if there is no source");
}, "audio.src - empty if no source");

test(function() {
 assert_equals(
  document.getElementById("video").src,
  "",
  "videoElement.src should be empty if there is no source");
}, "video.src - empty if no source");
  </script>
  </script>
 </body>
</html>
